+++ kinto_client/__init__.py
@@ -1,13 +1,12 @@
 import collections
-import requests
 import uuid
 from six import iteritems
-from six.moves.urllib.parse import urlparse
 
 from contextlib import contextmanager
 
 
 from kinto_client import utils
+from kinto_client.session import create_session, Session
 from kinto_client.batch import Batch
 from kinto_client.exceptions import BucketNotFound, KintoException
 
@@ -25,31 +24,6 @@ OBJECTS_PERMISSIONS = {
 
 ID_FIELD = 'id'
 DO_NOT_OVERWRITE = {'If-None-Match': '*'}
-
-
-def create_session(server_url=None, auth=None, session=None):
-    """Returns a session from the passed arguments.
-
-    :param server_url:
-        The URL of the server to use, with the prefix.
-    :param auth:
-        A requests authentication policy object.
-    :param session:
-        An optional session object to use, rather than creating a new one.
-    """
-    # XXX Refactor the create_session to take place in the caller objects.
-    # E.g. test if the session exists before calling create_session.
-    if session is not None and (
-            server_url is not None or auth is not None):
-        msg = ("You cannot specify session and server_url or auth. "
-               "Chose either session or (auth + server_url).")
-        raise AttributeError(msg)
-    if session is None and server_url is None and auth is None:
-        msg = ("You need to either set session or auth + server_url")
-        raise AttributeError(msg)
-    if session is None:
-        session = Session(server_url=server_url, auth=auth)
-    return session
 
 
 class Endpoints(object):
@@ -81,78 +55,35 @@ class Endpoints(object):
                                  field=','.join(e.args)))
 
 
-class Session(object):
-    """Handles all the interactions with the network.
-    """
-    def __init__(self, server_url, auth=None):
-        self.server_url = server_url
-        self.auth = auth
-
-    def request(self, method, endpoint, data=None, permissions=None,
-                payload=None, **kwargs):
-        parsed = urlparse(endpoint)
-        if not parsed.scheme:
-            actual_url = utils.urljoin(self.server_url, endpoint)
-        else:
-            actual_url = endpoint
-
-        if self.auth is not None:
-            kwargs.setdefault('auth', self.auth)
-
-        payload = payload or {}
-        # if data is not None:
-        payload['data'] = data or {}
-        if permissions is not None:
-            if hasattr(permissions, 'as_dict'):
-                permissions = permissions.as_dict()
-            payload['permissions'] = permissions
-        if payload:
-            payload_kwarg = 'data' if 'files' in kwargs else 'json'
-            kwargs.setdefault(payload_kwarg, payload)
-        resp = requests.request(method, actual_url, **kwargs)
-        if not (200 <= resp.status_code < 400):
-            message = '{0} - {1}'.format(resp.status_code, resp.json())
-            exception = KintoException(message)
-            exception.request = resp.request
-            exception.response = resp
-            raise exception
-
-        if resp.status_code == 304:
-            body = None
-        else:
-            body = resp.json()
-        # XXX Add the status code.
-        return body, resp.headers
-
-
 class Client(object):
 
     def __init__(self, server_url=None, session=None, auth=None,
-                 bucket="default", collection=None):
+                 bucket="default", collection=None, retry=0, retry_after=None):
         self.endpoints = Endpoints()
-        self.session = create_session(server_url, auth, session)
+        self.session_kwargs = dict(server_url=server_url,
+                                   auth=auth,
+                                   session=session,
+                                   retry=retry,
+                                   retry_after=retry_after)
+        self.session = create_session(**self.session_kwargs)
         self._bucket_name = bucket
         self._collection_name = collection
         self._server_settings = None
 
     def clone(self, **kwargs):
-        return Client(**{
-            'session': kwargs.get('session', self.session),
-            'bucket': kwargs.get('bucket', self._bucket_name),
-            'collection': kwargs.get('collection', self._collection_name),
-        })
+        kwargs.setdefault('session', self.session)
+        kwargs.setdefault('bucket', self._bucket_name)
+        kwargs.setdefault('collection', self._collection_name)
+        return Client(**kwargs)
 
     @contextmanager
-    def batch(self, retry=0, retry_after=None, **kwargs):
+    def batch(self, **kwargs):
         if self._server_settings is None:
             resp, _ = self.session.request("GET", self._get_endpoint('root'))
             self._server_settings = resp['settings']
 
         batch_max_requests = self._server_settings['batch_max_requests']
-        batch = Batch(self,
-                      batch_max_requests=batch_max_requests,
-                      retry=retry,
-                      retry_after=retry_after)
+        batch = Batch(self, batch_max_requests=batch_max_requests)
         yield self.clone(session=batch, **kwargs)
         for (resp, headers) in batch.send():
             for i, response in enumerate(resp['responses']):
