+++ kinto_client/tests/test_batch.py
@@ -1,7 +1,6 @@
 import unittest2 as unittest
 import mock
 
-from kinto_client import Client
 from kinto_client.batch import Batch
 
 
@@ -94,106 +93,3 @@ class BatchRequestsTest(unittest.TestCase):
         calls = self.client.session.request.call_args_list
         _, kwargs1 = calls[0]
         assert kwargs1['payload']['requests'][0]['path'] == '/foobar'
-
-
-class RetryBatchTest(unittest.TestCase):
-
-    def setUp(self):
-        self.client = Client('https://server.com')
-        patch = mock.patch('kinto_client.requests.request')
-        self.addCleanup(patch.stop)
-        self.request_mocked = patch.start()
-
-        self.response_200 = mock.MagicMock()
-        self.response_200.status_code = 200
-        self.response_200.json().return_value = mock.sentinel.resp,
-        self.response_200.headers = mock.sentinel.headers
-
-        body_503 = {
-            "message": "Service temporary unavailable due to overloading",
-            "code": 503,
-            "error": "Service Unavailable",
-            "errno": 201
-        }
-        headers_503 = {
-            "Content-Type": "application/json; charset=UTF-8",
-            "Content-Length": 151
-        }
-        self.response_503 = mock.MagicMock()
-        self.response_503.status_code = 503
-        self.response_503.json.return_value = body_503
-        self.response_503.headers = headers_503
-
-        self.request_mocked.side_effect = [self.response_503]
-
-    def test_does_not_retry_by_default(self):
-        batch = Batch(self.client)
-        batch.request('GET', '/v1/foobar')
-        with self.assertRaises(Exception):
-            batch.send()
-
-    def test_succeeds_on_retry(self):
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_200]  # retry 1
-
-        batch = Batch(self.client, retry=1)
-        batch.request('GET', '/v1/foobar')
-        batch.send()
-
-    def test_can_retry_several_times(self):
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_503,  # retry 1
-                                           self.response_200]  # retry 2
-
-        batch = Batch(self.client, retry=2)
-        batch.request('GET', '/v1/foobar')
-        batch.send()
-
-    def test_fails_if_retry_exhausted(self):
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_503,  # retry 1
-                                           self.response_503,  # retry 2
-                                           self.response_200]
-
-        batch = Batch(self.client, retry=2)
-        batch.request('GET', '/v1/foobar')
-        with self.assertRaises(Exception):
-            batch.send()
-
-    def test_does_not_wait_if_retry_after_header_is_not_present(self):
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_200]
-        with mock.patch('kinto_client.batch.time.sleep') as sleep_mocked:
-            batch = Batch(self.client, retry=1)
-            batch.request('GET', '/v1/foobar')
-            batch.send()
-            sleep_mocked.assert_called_with(0)
-
-    def test_waits_if_retry_after_header_is_present(self):
-        self.response_503.headers["Retry-After"] = 27
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_200]
-        with mock.patch('kinto_client.batch.time.sleep') as sleep_mocked:
-            batch = Batch(self.client, retry=1)
-            batch.request('GET', '/v1/foobar')
-            batch.send()
-            self.assertTrue(sleep_mocked.called)
-
-    def test_waits_if_retry_after_is_forced(self):
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_200]
-        with mock.patch('kinto_client.batch.time.sleep') as sleep_mocked:
-            batch = Batch(self.client, retry=1, retry_after=10)
-            batch.request('GET', '/v1/foobar')
-            batch.send()
-            sleep_mocked.assert_called_with(10)
-
-    def test_forced_retry_after_overrides_value_of_header(self):
-        self.response_503.headers["Retry-After"] = 27
-        self.request_mocked.side_effect = [self.response_503,
-                                           self.response_200]
-        with mock.patch('kinto_client.batch.time.sleep') as sleep_mocked:
-            batch = Batch(self.client, retry=1, retry_after=10)
-            batch.request('GET', '/v1/foobar')
-            batch.send()
-            sleep_mocked.assert_called_with(10)
