+++ kinto_client/tests/test_session.py
@@ -1,12 +1,14 @@
 import mock
 
+from kinto_client.session import Session, create_session
+from kinto_client.exceptions import KintoException
+
 from .support import unittest
-from kinto_client import (Session, create_session, KintoException)
 
 
 class SessionTest(unittest.TestCase):
     def setUp(self):
-        p = mock.patch('kinto_client.requests')
+        p = mock.patch('kinto_client.session.requests')
         self.requests_mock = p.start()
         self.addCleanup(p.stop)
 
@@ -117,14 +119,16 @@ class SessionTest(unittest.TestCase):
     def test_initialization_fails_on_missing_args(self):
         self.assertRaises(AttributeError, create_session)
 
-    @mock.patch('kinto_client.Session')
+    @mock.patch('kinto_client.session.Session')
     def test_creates_a_session_if_needed(self, session_mock):
         # Mock the session response.
         create_session(server_url=mock.sentinel.server_url,
                        auth=mock.sentinel.auth)
         session_mock.assert_called_with(
             server_url=mock.sentinel.server_url,
-            auth=mock.sentinel.auth)
+            auth=mock.sentinel.auth,
+            retry=0,
+            retry_after=None)
 
     def test_use_given_session_if_provided(self):
         session = create_session(session=mock.sentinel.session)
@@ -137,3 +141,94 @@ class SessionTest(unittest.TestCase):
         session = Session('https://example.org')
         body, headers = session.request('get', 'https://example.org/test')
         assert body is None
+
+
+class RetryRequestTest(unittest.TestCase):
+
+    def setUp(self):
+        p = mock.patch('kinto_client.session.requests')
+        self.requests_mock = p.start()
+        self.addCleanup(p.stop)
+
+        self.response_200 = mock.MagicMock()
+        self.response_200.status_code = 200
+        self.response_200.json().return_value = mock.sentinel.resp,
+        self.response_200.headers = mock.sentinel.headers
+
+        body_503 = {
+            "message": "Service temporary unavailable due to overloading",
+            "code": 503,
+            "error": "Service Unavailable",
+            "errno": 201
+        }
+        headers_503 = {
+            "Content-Type": "application/json; charset=UTF-8",
+            "Content-Length": 151
+        }
+        self.response_503 = mock.MagicMock()
+        self.response_503.status_code = 503
+        self.response_503.json.return_value = body_503
+        self.response_503.headers = headers_503
+
+        self.requests_mock.request.side_effect = [self.response_503]
+
+    def test_does_not_retry_by_default(self):
+        session = Session('https://example.org')
+        with self.assertRaises(KintoException):
+            session.request('GET', '/v1/foobar')
+
+    def test_succeeds_on_retry(self):
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_200]  # retry 1
+        session = Session('https://example.org', retry=1)
+        session.request('GET', '/v1/foobar')  # Not raising.
+
+    def test_can_retry_several_times(self):
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_503,  # retry 1
+                                                  self.response_200]  # retry 2
+        session = Session('https://example.org', retry=2)
+        session.request('GET', '/v1/foobar')  # Not raising.
+
+    def test_fails_if_retry_exhausted(self):
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_503,  # retry 1
+                                                  self.response_503,  # retry 2
+                                                  self.response_200]
+        session = Session('https://example.org', retry=2)
+        with self.assertRaises(KintoException):
+            session.request('GET', '/v1/foobar')
+
+    def test_does_not_wait_if_retry_after_header_is_not_present(self):
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_200]
+        with mock.patch('kinto_client.session.time.sleep') as sleep_mocked:
+            session = Session('https://example.org', retry=1)
+            session.request('GET', '/v1/foobar')
+            sleep_mocked.assert_called_with(0)
+
+    def test_waits_if_retry_after_header_is_present(self):
+        self.response_503.headers["Retry-After"] = 27
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_200]
+        with mock.patch('kinto_client.session.time.sleep') as sleep_mocked:
+            session = Session('https://example.org', retry=1)
+            session.request('GET', '/v1/foobar')
+            self.assertTrue(sleep_mocked.called)
+
+    def test_waits_if_retry_after_is_forced(self):
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_200]
+        with mock.patch('kinto_client.session.time.sleep') as sleep_mocked:
+            session = Session('https://example.org', retry=1, retry_after=10)
+            session.request('GET', '/v1/foobar')
+            sleep_mocked.assert_called_with(10)
+
+    def test_forced_retry_after_overrides_value_of_header(self):
+        self.response_503.headers["Retry-After"] = 27
+        self.requests_mock.request.side_effect = [self.response_503,
+                                                  self.response_200]
+        with mock.patch('kinto_client.session.time.sleep') as sleep_mocked:
+            session = Session('https://example.org', retry=1, retry_after=10)
+            session.request('GET', '/v1/foobar')
+            sleep_mocked.assert_called_with(10)
